+++ src/Builder/BuilderCollection.php
@@ -15,7 +15,6 @@ declare(strict_types=1);
 namespace Ramsey\Uuid\Builder;
 
 use Ramsey\Collection\AbstractCollection;
-use Ramsey\Collection\CollectionInterface;
 use Ramsey\Uuid\Converter\Number\GenericNumberConverter;
 use Ramsey\Uuid\Converter\Time\GenericTimeConverter;
 use Ramsey\Uuid\Converter\Time\PhpTimeConverter;
@@ -28,7 +27,7 @@ use Traversable;
 /**
  * A collection of UuidBuilderInterface objects
  */
-class BuilderCollection extends AbstractCollection implements CollectionInterface
+class BuilderCollection extends AbstractCollection
 {
     public function getType(): string
     {