+++ src/Nonstandard/UuidV6.php
@@ -21,6 +21,7 @@ use Ramsey\Uuid\Converter\NumberConverterInterface;
 use Ramsey\Uuid\Converter\TimeConverterInterface;
 use Ramsey\Uuid\Exception\DateTimeException;
 use Ramsey\Uuid\Exception\InvalidArgumentException;
+use Ramsey\Uuid\Lazy\LazyUuidFromString;
 use Ramsey\Uuid\Rfc4122\FieldsInterface as Rfc4122FieldsInterface;
 use Ramsey\Uuid\Rfc4122\UuidInterface;
 use Ramsey\Uuid\Rfc4122\UuidV1;
@@ -106,10 +107,10 @@ final class UuidV6 extends Uuid implements UuidInterface
             . '1' . substr($hex, 0, 3)
             . substr($hex, 16);
 
-        /** @var UuidV1 $uuid */
+        /** @var LazyUuidFromString $uuid */
         $uuid = Uuid::fromBytes((string) hex2bin($hex));
 
-        return $uuid;
+        return $uuid->toUuidV1();
     }
 
     /**
@@ -124,9 +125,9 @@ final class UuidV6 extends Uuid implements UuidInterface
             . '6' . substr($hex, 5, 3)
             . substr($hex, 16);
 
-        /** @var UuidV6 $uuid */
+        /** @var LazyUuidFromString $uuid */
         $uuid = Uuid::fromBytes((string) hex2bin($hex));
 
-        return $uuid;
+        return $uuid->toUuidV6();
     }
 }