+++ tests/TestCase.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace Ramsey\Uuid\Test;
 
-use AspectMock\Test as AspectMock;
 use Mockery;
 use PHPUnit\Framework\TestCase as PhpUnitTestCase;
 
@@ -17,12 +16,14 @@ class TestCase extends PhpUnitTestCase
     protected function tearDown(): void
     {
         parent::tearDown();
-        AspectMock::clean();
         Mockery::close();
     }
 
     public static function isLittleEndianSystem(): bool
     {
-        return current(unpack('v', pack('S', 0x00FF))) === 0x00FF;
+        /** @var array $unpacked */
+        $unpacked = unpack('v', pack('S', 0x00FF));
+
+        return current($unpacked) === 0x00FF;
     }
 }